move more utility methods
